Optionally store a previously initialized client
The NUOPC driver initializes all the SmartRedis clients for use throughout
CESM. To avoid duplicating initialization efforts, smartredis_init was
rewritten to simply store this client within the MOM control structure if
an optional argument is passed to initialize_MOM. Additionally, a check
is included to ensure that this client is indeed initialized if the user
had set USE_SMARTSIM = True.

This commit also updates the SmartRedis stub in the external directory
to reflect changes made by @jedwards4b.